Add files via upload
If you are snooping around on this depository and try to run this program, let me know if you cannot run the files, or if something doesn't work. I may have to change the file path because I put the CSS file and java files in index folders, and it wouldn't let me upload the folders.